+++ CHANGELOG.md
@@ -22,6 +22,7 @@ This project <em>does not yet</em> adhere to [Semantic Versioning](https://semve
 - DwC Import: Disabled and Ignored cells have the same color
 - New taxon name: The scroll bar is missing in the right column [#4932]
 - Sometimes, the autocomplete spinner doesn't appear [#4946]
+- Reset sort direction when sorting by a different column in filter table
 
 [#4932]: https://github.com/SpeciesFileGroup/taxonworks/issues/4932
 [#4946]: https://github.com/SpeciesFileGroup/taxonworks/issues/4946